requirements.txt
torchserve
torch-model-archiver
torchvision
Pillow
